backend/pyproject.toml
[project]
name = "crimescope-backend"
version = "1.3.0"
description = "CrimeScope Swarm Intelligence Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "supabase>=2.0.0",
    "neo4j>=5.14.0",
    "chromadb>=0.4.18",
    "mem0ai>=0.1.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "python-docx>=1.0.0",
    "PyMuPDF>=1.23.0",
    "moviepy>=1.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
